Separate warnings from main options for RGBLINK
